+++ packages/tokens/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [@synergy-design-system/tokens-v1.0.0-main.4](https://github.com/SickDesignSystem/synergy/compare/tokens/1.0.0-main.3...tokens/1.0.0-main.4) (2023-11-02)
+
+
+### Features
+
+* ✨ add button tokens ([#109](https://github.com/SickDesignSystem/synergy/issues/109)) ([237148c](https://github.com/SickDesignSystem/synergy/commit/237148ccf64be74fc3721dc5381010c5495890f8))
+
 # [@synergy-design-system/tokens-v1.0.0-main.3](https://github.com/SickDesignSystem/synergy/compare/tokens/1.0.0-main.2...tokens/1.0.0-main.3) (2023-10-26)